\name{calQuadrat}
\alias{calQuadrat}
\title{Quadrat d'un vector d'elements}
\usage{
  calQuadrat(x)
}
\arguments{
  \item{x}{vector d'elements}
}
\value{
  etorna el quadrat del vector d'elements.
}
\description{
  calQuadrat retorna el quadrat d'un vector de tots els
  valors presents al seus arguments.
}
\examples{
calQuadrat(10)
calQuadrat(1:10)
}
\author{
  Lluis Ramon
}
\references{
  RUGBCN
}

